added status bar, toString
