+++ src/by/itacademy/hw10/task4/Main.java
@@ -5,8 +5,9 @@ import by.itacademy.hw10.task4.exception.AnimalNotExistException;
 import by.itacademy.hw10.task4.model.Gender;
 import by.itacademy.hw10.task4.model.animal.Cat;
 import by.itacademy.hw10.task4.model.animal.Dog;
-import by.itacademy.hw10.task4.service.AnimalJournalHandler;
-import by.itacademy.hw10.task4.view.AnimalJournalView;
+import by.itacademy.hw10.task4.service.AnimalJournalHandlerImpl;
+import by.itacademy.hw10.task4.view.AnimalJournalViewImpl;
+import by.itacademy.hw10.task4.view.StatusBar;
 
 import java.util.Calendar;
 import java.util.GregorianCalendar;
@@ -15,8 +16,8 @@ public class Main {
 
     public static void main(String[] args) {
 
-        AnimalJournalHandler handler = new AnimalJournalHandler();
-        AnimalJournalView journalView = new AnimalJournalView(handler);
+        AnimalJournalHandlerImpl handler = new AnimalJournalHandlerImpl(new StatusBar());
+        AnimalJournalViewImpl journalView = new AnimalJournalViewImpl(handler);
 
         journalView.printAvailableAnimals();
 
@@ -50,17 +51,15 @@ public class Main {
             journalView.printException(e);
         }
 
-        System.out.println("Сортируем по цене");
         journalView.print(handler.sortByPrice());
 
-        System.out.println("Сортируем по кличке");
         journalView.print(handler.sortByName());
 
-        System.out.println("Сортируем по породе и дате рождения");
         journalView.print(handler.sortByBreedAndDateOfBirth());
 
-        System.out.println("Поиск по цене от 0 до 60:");
         journalView.print(handler.searchAnimalByPrice(0, 60));
+
+        journalView.printSoldAnimals();
     }
 
 }
