+++ src/by/itacademy/hw10/task4/exception/AnimalException.java
@@ -1,0 +1,9 @@
+package by.itacademy.hw10.task4.exception;
+
+public class AnimalException extends Exception{
+
+    public AnimalException(String message) {
+        super(message);
+    }
+
+}
