+++ src/by/itacademy/hw10/task4/exception/AnimalNotExistException.java
@@ -1,6 +1,6 @@
 package by.itacademy.hw10.task4.exception;
 
-public class AnimalNotExistException extends Exception{
+public class AnimalNotExistException extends AnimalException{
 
     public AnimalNotExistException(String message) {
         super(message);
