+++ src/by/itacademy/hw10/task4/view/StatusBar.java
@@ -1,0 +1,7 @@
+package by.itacademy.hw10.task4.view;
+
+public class StatusBar {
+    public void log(String message){
+        System.out.println(message);
+    }
+}
